requirements.txt
numpy==1.26.3
openai==1.10.0
#openapi-schema-pydantic==1.2.4
pandas==2.1.4
pinecone-client==2.2.4
streamlit==1.31.0
streamlit-chat==0.1.1
tiktoken==0.5.2
cassandra-driver
cassio>=0.1.1
gradientai
llama-index>=0.5.0
pypdf
llama-index-vector-stores-cassandra
llama-index-llms-gradient
llama-index-embeddings-gradient
gradio>=4.19
python-dotenv
langchain
astrapy
fitz
frontend